requirements.txt
tk
tabulate
numpy
seaborn
matplotlib
python-docx
pymongo
flask
flask-wtf
wtforms
flask_pymongo
scikit-learn
reportlab
pdfpages
opencv-python
Pillow
ipython
docx2pdf
requests